fix: handle ctrl-c in a consistent way, always aborting
Ctrl-C now always raise KeyboardInterrupt no matter what kind of prompt is used.
This is currently handled by the tour by printing "Aborted." and exiting, but
it's also ready for our future plans of saving the partial state on interrupt,
which will be able to do it consistently with a try/except KeyboardInterrupt block.

Fixes #85