fix: quiet duplicate text warning during attn processing
The text warnings are displayed while processing the text, and again while
processing attn. Text processing is very fast so it's not really an issue not
to see the progress bar well, whereas attn is slow, so the solution here is
simply to suppress the warnings during attn processing.

Fixes #70