feat: display a summary of missing symbols after processing "text"
Disable parallel processing for the "text" step, which is OK since it gains
very little, in order for the missing_symbols counters to get updated correctly
in the main process, and thus allow us to print it, instead of dumping a bump
of messages every time a symbol is ignored.

Fixes #70 even better than the previous commit. :)