feat: replace repeated prepro warnings by a summary of missing symbols
Squashes for following commits:

* test: run_test now runs the dev suite by default

* fix: quiet duplicate text warning during attn processing

The text warnings are displayed while processing the text, and again while
processing attn. Text processing is very fast so it's not really an issue not
to see the progress bar well, whereas attn is slow, so the solution here is
simply to suppress the warnings during attn processing.

Fixes #70

* feat: display a summary of missing symbols after processing "text"

Disable parallel processing for the "text" step, which is OK since it gains
very little, in order for the missing_symbols counters to get updated correctly
in the main process, and thus allow us to print it, instead of dumping a bump
of messages every time a symbol is ignored.

Fixes #70 even better than the previous commit. :)

* test: improve PR patch coverage

* fix: pfs also needs to be treated like text for warnings

* fix: make sure the pfs and text missing counters are distinct

* fix: grammatically correct messages, and doctest for everyvoce.utils