refactor: remove unreachable code and bump coverage
reload_filelist_data_as_dict() can't have to reload a festival file, so replace
the unreachable code by an assertion.

Also bump coverage by using the null_patch() and HasLanguage and HasSpeaker
when testing festival, so that we run the prompt() function and let it detect
that the format is festival.