ci: up our security stance with pinned actions (#753)
Triggered by the supply chain hihacking of tj-actions/changed-files
See
 - https://www.stepsecurity.io/blog/harden-runner-detection-tj-actions-changed-files-action-is-compromised
 - https://www.stepsecurity.io/blog/reviewdog-github-actions-are-compromised
 - https://github.blog/changelog/2025-08-15-github-actions-policy-now-supports-blocking-and-sha-pinning-actions/
 - https://github.blog/changelog/2025-08-26-releases-now-support-immutability-in-public-preview/

There are several more changes we could make to harden our security stance, but
I stopped at what's easy and not too disruptive. It's a wake up call that an
action we use got compromised and we could have been affected if we had run
workflows at the wrong time.